+++ src/psa_attestation.c
@@ -21,6 +21,8 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  */
 
+#include "psa_config.h"
+
 #include <stddef.h>
 #include <stdint.h>